refactor to put in same file (savings of 2.5 seconds)
